Add files via upload
The Blood Management System is a comprehensive web application that plays a pivotal role in efficiently managing blood donations and requests. Developed using a stack comprising HTML, CSS, JavaScript, and PHP, this system addresses critical needs within the healthcare sector. It streamlines the entire process of blood donation, making it easier for donors to contribute and for medical institutions to access lifesaving blood units when needed.